Cleaned up layer violations
