+++ front-end/src/hooks/WasmStateInit.ts
@@ -7,8 +7,6 @@ import {
 import { Observable, Subject } from 'rxjs';
 import { SessionController } from './SessionController';
 
-export type GameHexes = Record<string, string>;
-
 var chia_gaming_init: WasmInitFn | undefined = undefined;
 var cg: WasmConnection | undefined = undefined;
 var logInitialized = false;
@@ -145,19 +143,7 @@ export class WasmStateInit {
     // return this.wasmConnection?.chia_identity(rngSeed);
   }
 
-  async loadGameHexes(): Promise<GameHexes> {
-    const dec = new TextDecoder();
-    const fetchText = (url: string) => this.fetchPreset(url).then(b => dec.decode(b));
-    const [calpoker, spacepoker, krunk] = await Promise.all([
-      fetchText('clsp/games/calpoker/calpoker_include_calpoker_factory.hex'),
-      fetchText('clsp/games/spacepoker/spacepoker_include_spacepoker_factory.hex'),
-      fetchText('clsp/games/krunk/krunk_include_krunk_factory.hex'),
-    ]);
-    return { calpoker, spacepoker, krunk };
-  }
-
   createGame(
-    gameHexes: GameHexes,
     rngId: number,
     wasm: WasmConnection,
     have_potato: boolean,
@@ -167,13 +153,8 @@ export class WasmStateInit {
     channelTimeout = 15,
     unrollTimeout = 15,
   ): { game: ChiaGame, puzzleHash: string } {
-    const game_types: Record<string, { version: number; hex: string }> = {};
-    for (const [name, hex] of Object.entries(gameHexes)) {
-      game_types[name] = { version: 1, hex };
-    }
     const result = wasm.create_game_session({
       rng_id: rngId,
-      game_types,
       have_potato: have_potato,
       my_contribution: { amt: my_contribution },
       their_contribution: { amt: their_contribution },
@@ -198,13 +179,4 @@ export class WasmStateInit {
     let chiaGameId = wasm.restore_session(serializedGame, seedHex);
     return new ChiaGame(wasm, chiaGameId);
   }
-}
-
-export async function loadGameHexes(fetchHex: (filename: string) => Promise<string>): Promise<GameHexes> {
-  const [calpoker, spacepoker, krunk] = await Promise.all([
-    fetchHex('clsp/games/calpoker/calpoker_include_calpoker_factory.hex'),
-    fetchHex('clsp/games/spacepoker/spacepoker_include_spacepoker_factory.hex'),
-    fetchHex('clsp/games/krunk/krunk_include_krunk_factory.hex'),
-  ]);
-  return { calpoker, spacepoker, krunk };
 }
